moved isEmbed to hooks utils package, updated docs
