+++ apps/cat/src/App.tsx
@@ -1,5 +1,5 @@
-import { Suspense, useMemo, useState } from "react";
-import { BrowserRouter, Routes, Route, useLocation, useNavigate } from "react-router-dom";
+import { Suspense, useState } from "react";
+import { BrowserRouter, Routes, Route, useNavigate } from "react-router-dom";
 import Skeleton from "@mui/material/Skeleton";
 import Box from "@mui/material/Box";
 import { useTranslation } from "react-i18next";
@@ -24,18 +24,7 @@ import {
   FacetedWrapper,
   FacetedSearchProvider,
 } from "@dans-framework/rdt-search-ui";
-
-const useEmbedHandler = () => {
-  const location = useLocation();
-
-  // on first render, check if embed=true is present in the url
-  const isEmbed = useMemo(() => {
-    const queryParams = new URLSearchParams(location.search);
-    return queryParams.get("embed") === "true";
-  }, []);
-
-  return { isEmbed };
-};
+import { useEmbedHandler } from "@dans-framework/utils";
 
 const App = () => {
   const { i18n } = useTranslation();
