+++ packages/utils/hooks/package.json
@@ -1,0 +1,16 @@
+{
+    "name": "@dans-framework/utils/hooks",
+    "main": "index.ts",
+    "version": "1.0.0",
+    "description": "",
+    "keywords": [],
+    "author": "",
+    "dependencies": {
+        "react": "^18.2.0",
+        "react-router-dom": "^6.14.2"
+    },
+    "devDependencies": {
+      "@types/react": "^18.2.15"
+    }
+  }
+
