+++ packages/utils/lib/index.ts
@@ -4,3 +4,4 @@ export * from "../error";
 export * from "../typescript";
 export * from "../preloader";
 export * from "../user";
+export * from "../hooks";
